SpaczzRuler compatible with spacy v2/v3.
